Create Knapsack.cpp
This algorithm works by storing the solutions to the subproblems in a table. The algorithm then iteratively solves the larger problems by using the solutions to the subproblems that have already been solved.

The knapsack problem is a NP-hard problem, meaning that there is no known polynomial-time algorithm for solving it. However, the dynamic programming algorithm shown above is a relatively efficient algorithm for solving the knapsack problem.