Create Pangram.js
This JavaScript code defines a function called isPangram that checks if a given input string contains all the letters of the English alphabet, regardless of their case (i.e., it's case-insensitive). It iterates through the alphabet and checks if each letter is present in the input string. If any letter is missing, it returns false; otherwise, it returns true to indicate that the input is a pangram, a sentence containing all the letters of the alphabet.